+++ docs/docs/fireback-workspaces/querydsl.mdx
@@ -8,3 +8,6 @@ in text format, which Fireback has incorporated.
 
 QuerDSL struct is super important, and is available in almost every
 function related to Fireback as an argument.
+
+You can assume that `QueryDSL` is in very similar to Context argument
+in other frameworks.
